tui: refactor to accept commit history
Preparation for fetching full commit history from PR, but for now stays
compatible with the current solution of providing from and to commits.
Arguments are now provided as flags instead of positional, and the
program no longer accepts them being empty, but they are not required
by serde. This is because once PR history can be fetched, they can be
left empty again, since they'll just be set from first and last commit
from the PR history by default.

Moved branch diff calculation out of the setters into the event loop,
which is why we're storing the indexes now to detect changes and
recalculate if needed.