tui: add keybinds to move between patchset versions
Any arbitrary base should be able to be compared to a commit that comes
at least after it. Allowing base and comparison to be equal might be
slightly confusing at first, but this will work out just fine and render
the commit's diff by itself compared to its parent without any rebasing.