prepare for fetching PR history from git forge
Introduce trait for parsing a PR URL, and fetching the history from
whatever git forge the URL points to. Might be fallible if the forge is
on a custom host that does not include the service's name directly, so
maybe a CLI arg to pass forge type might be useful in the future.