support rendering diffs for added and removed commits
Refactors diff rendering to use intermediate DiffTree, so that
calculation of the diff (including optional rebasing for interdiff) and
actual rendering are separated.
AddedCommit can simply compare against its parent tree, and to render
the diff of a removed commit correctly, reverse it and diff its parent
tree against the commit that was removed.