fetch missing refs on startup
Ensures that all diffs can be generated by getting the list of refs
that was returned from a PR history, and checking if all of them exist.
If not, fetch the missing ones to <sha>:refs/remotes/<remote>/<sha>, so
that jj will import them.
Important is that the jj import will not update the existing repo, but
return a new one after committing the transaction, so we'll need to
update repo after running this.

Some slight refactoring to only init the App struct after this has been
done might be needed, and moving loading the repo from the workspace out
of its initialization.