diff: fix diff calculation for partial rebases
If only half the commits in a branch are updated (e.g. when squashing
something into a commit in the middle), then the unchanged commits were
not handled correctly but shown as new every time.
Fix this by updating the fork_point revset to also include trunk, so
that these commits will not appear.

Commit stat calculation and UI styles also need to be handled for these
cases, so if we have a from and to, but the shas are identical, this can
be handled as that new case.