add support for copying diff to clipboard
Viewing a plain git diff (even with the line coloring) isn't always the
nicest, but to keep it simple, copying to clipboard allows viewing in
any other diff viewer, most of those should support git style diffs.

The main use of this tool should be generating the interdiffs in the
first place, since that's where the other tools (like Bitbucket, Github
etc) fail, even if they're nicer for viewing them.