rewrite TUI with threaded event model
Model uses three threads, one main for coordination, one for rendering
UI and taking input (this one has a exclusive terminal access, which is
why these are on the same) and one worker thread to delegate heavy tasks
to, such as calculating diffs or fetching more pages.

Main benefit is pagination support for the PR fetcher trait, since
bitbucket by default limits to 25, and does not filter for rescope
events on a request level.

UI rendering is stateless and happens via screen snapshots sent from the
main thread, which for now clones the current screen data. If this needs
more optimization, Arcs could be introduced here.

All the pre-rewrite functionality is present again, so this is a good
point to ship it and improve on it later when needed.