PrFetcher: introduce cursor based pagination support
Refactors pagination to handle different pagination types. The fetcher
itself checks if it's being handed the correct type at runtime, but
because the initial pagination is None and after that has always been
returned by the fetcher itself, this is an acceptable tradeoff.

This is in preparation for introducing GitHub support via GraphQL which
is different from the previous abstraction in that it uses cursor based
pagination and returns the oldest events first, unlike bitbucket.