worker: print missing commits message as app screen
Instead of using println which destroys the app layout, it now sends it
as the new Loading Response type which was introduced.
ensure_commits_exist, and fetch_commits were split, so that the number
of missing commits can be reported before fetching them.
Important to note, the appscreen transitions have a special exception
for receiving LoadCommits while on Loading, to initialize the first
branch diff calculation, so in order to not introduce another screen
type into this exception, loading was repurposed to display custom
loading messages as well.