add support for plain git repos
This does not mean the tool now works without jj, it still relies on it
for change tracking, rebasing and diff rendering, but now the repo it is
used on no longer needs to have a jj repo initialized, and also won't be
modified.
Instead, now a tempdir is created, a jj repo with the workspace pointing
to the active git repo is created there, and then this is used in
further operation just the same as an already existing jj repo.

Because the tempdir crate deletes the directory on disk on drop, the
handle needs to return it so it will only be dropped once the actual
application exits.