PrFetcher: add warning if token is unset
Will only be logged on startup, so not the ideal place since it will be
hidden by the TUI immediately, but at least if fetching runs into an
error, after exiting users will see the likely cause of the issue.
Token setup is a one time thing, so this should be fine for the time
being.