fix trunk detection when importing git repo
Only user settings are retained in memory, since we're not actually
persisting the workspace config on disk. So the trunk alias was
defaulting to another value, which does work on most repos using master
or main, but did not work with a trunk named develop.

This fixes it by writing the alias into user settings, which we keep if
we reuse the repo that is initialized, and uses the correct name for the
alias including the brackets.