fix relative path resolution when importing git repo
gix would open a repo correctly when using a relative path, but then
somewhere down the line the path was not resolved correctly, leading the
temporary jj repo to point to an invalid git repository. This is fixed
by using absolute paths when opening the git repository.

Opening the git repo was also moved before initializing the temp dir
because there was no point in creating and immediately deleting it again
in case opening already failed.

Fun to debug when testing only with `cargo run -- --repo
<some-absolute-path>` in debug mode, and then only ever running the
installed release version from inside a git repo, where of course the
default `.` is also a relative path which would fail to resolve.