mark commits as changed when only the message is updated
The UI was already displaying the commit with a marker for commit
message changes, but because `has_changes` returned false for them, the
default filtered view would not show commits with only message but not
code changes by default. Because this is just as relevant for review,
these will now not be hidden anymore.